branch_lifecycle

+++ composeApp/src/commonMain/kotlin/ru/llama/tool/data/room/ai_properties_db/AiPropertiesEntity.kt
@@ -7,7 +7,7 @@ import ru.llama.tool.data.room.DBConst
 
 @Entity(tableName = DBConst.PROPERTIES_TABLE)
 data class AiPropertiesEntity(
-    @PrimaryKey(autoGenerate = false) @ColumnInfo(name = "id") val id: Int,
+    @PrimaryKey(autoGenerate = false) @ColumnInfo(name = "id") val id: Long,
     @ColumnInfo(name = "system_prompt") val systemPrompt: String,
     @ColumnInfo(name = "temperature") val temperature: Double,
     @ColumnInfo(name = "max_tokens") val maxTokens: Int,
